1. Added Profile Search Fragment. 1.1. Added Games Search Fragment. 1.2. Added Heroes Search Fragment.

+++ app/src/main/java/com/ez/dotarate/model/repository/UserRepository.kt
@@ -18,6 +18,6 @@ interface UserRepository {
     fun getUserName(): LiveData<UserName?>
 
     // Network
-    suspend fun getUserResponse(id: Long): Response<UserResponse>
-    suspend fun getWinsAndLosses(id: Long): Response<WinsAndLosses>
+    suspend fun getUserResponse(id: Int): Response<UserResponse>
+    suspend fun getWinsAndLosses(id: Int): Response<WinsAndLosses>
 }
