+++ app/src/main/java/com/ez/dotarate/model/repository/UserRepositoryImpl.kt
@@ -76,7 +76,7 @@ constructor(
      * We don’t need to call enqueue() and implement callbacks anymore!
      * But notice, now our repo method is suspend too and returns a Response<UserResponse> object.
      */
-    override suspend fun getUserResponse(id: Long): Response<UserResponse> {
+    override suspend fun getUserResponse(id: Int): Response<UserResponse> {
         Log.d("MyLogs", "ПОШЁЛ ЗАПРОС ПОЛЬЗОВАТЕЛЯ. ID = $id")
         return api.getUser(id)
     }
@@ -86,7 +86,7 @@ constructor(
      * We don’t need to call enqueue() and implement callbacks anymore!
      * But notice, now our repo method is suspend too and returns a Response<WinsAndLosses> object.
      */
-    override suspend fun getWinsAndLosses(id: Long): Response<WinsAndLosses> {
+    override suspend fun getWinsAndLosses(id: Int): Response<WinsAndLosses> {
         Log.d("MyLogs", "ПОШЁЛ ЗАПРОС ВИНОВ И ЛУЗОВ. ID = $id")
         return api.getWinsAndLosses(id)
     }
